Move the config interface into the factory show make clear what the relation here is.
