--- Corona-Warn-App/src/main/java/de/rki/coronawarnapp/task/TaskConfig.kt
@@ -1,11 +1,0 @@
-package de.rki.coronawarnapp.task
-
-interface TaskConfig {
-
-    val collisionBehavior: CollisionBehavior
-
-    enum class CollisionBehavior {
-        ENQUEUE,
-        SKIP_IF_ALREADY_RUNNING
-    }
-}
