--- Corona-Warn-App/src/main/java/de/rki/coronawarnapp/task/TaskType.kt
@@ -1,5 +1,0 @@
-package de.rki.coronawarnapp.task
-
-enum class TaskType {
-    EXAMPLE
-}
